refactor: convert layout processor into a plugin
